feat(#1566): card-only welcome credits; unfunded generation through references
SMS was letting credit farmers claim the welcome grant. Drop Twilio
verification and replace that link with Ask the founder.

Skipping the grant still lets a team enhance and generate through
references — those stages are cheap enough to eat. Images, motion, and
music still require a card or credits.
