+++ .env.example
@@ -102,14 +102,6 @@ BETTER_AUTH_SECRET=
 # STRIPE_PUBLISHABLE_KEY=          # pk_test_... / pk_live_... for in-app card entry
 # STRIPE_WEBHOOK_SECRET=
 
-# Welcome credits by SMS (#1539, optional). Lets a user who cannot save a card
-# (WeChat Pay etc.) unlock the welcome grant by verifying a mobile number
-# instead. Needs a Twilio Verify service: https://console.twilio.com/us1/develop/verify/services
-# All three must be set; the claim dialog hides the SMS option otherwise.
-# TWILIO_ACCOUNT_SID=
-# TWILIO_AUTH_TOKEN=
-# TWILIO_VERIFY_SERVICE_SID=
-
 # BYOK API Key Encryption (required if users bring their own AI keys)
 # Auto-generated by `bun dev` in .env.local
 # API_KEY_ENCRYPTION_KEY=
