+++ scripts/push-secrets.ts
@@ -89,9 +89,6 @@ export const SECRETS = {
   STRIPE_PUBLISHABLE_KEY: { runtime: true, build: false },
   STRIPE_SECRET_KEY: { runtime: true, build: false },
   STRIPE_WEBHOOK_SECRET: { runtime: true, build: false },
-  TWILIO_ACCOUNT_SID: { runtime: true, build: false },
-  TWILIO_AUTH_TOKEN: { runtime: true, build: false },
-  TWILIO_VERIFY_SERVICE_SID: { runtime: true, build: false },
   VITE_APP_NAME: { runtime: true, build: true },
   VITE_APP_URL: { runtime: true, build: true },
   VITE_PUBLIC_POSTHOG_HOST: { runtime: true, build: true },
