+++ src/billing/ui/welcome-credits-dialog.tsx
@@ -2,8 +2,7 @@
  * Welcome Credits Dialog (#1096, #1516)
  *
  * - **claim**: Stripe on, $20 unpaid. Add a card (Stripe Checkout setup,
- *   no charge) to unlock it — or verify a mobile number by SMS when Twilio
- *   is configured (#1539; wallets like WeChat Pay cannot be saved as cards).
+ *   no charge) to unlock it. No card? Ask the founder (#1566).
  * - **gift**: unused signup grant and Stripe off (e2e / self-host).
  *
  * Dismiss cadence lives in localStorage (house pattern for UI prefs).
@@ -11,8 +10,6 @@
  */
 
 import { Button } from '@/ui/shadcn/button';
-import { Input } from '@/ui/shadcn/input';
-import { PhoneInput } from '@/ui/phone-input';
 import {
   Dialog,
   DialogContent,
@@ -25,9 +22,8 @@ import { Switch } from '@/ui/shadcn/switch';
 import {
   claimWelcomeCreditsFn,
   createSetupCheckoutSessionFn,
-  sendWelcomePhoneCodeFn,
-  verifyWelcomePhoneCodeFn,
 } from '@/billing/billing.fn';
+import { AskFounderCard } from '@/billing/ui/ask-founder-card';
 import {
   BILLING_BALANCE_KEY,
   BILLING_PAYMENT_METHODS_KEY,
@@ -41,7 +37,6 @@ import type { WelcomeDialogMode } from '@/billing/constants';
 import { microsToDisplayUsd } from '@/billing/money';
 import { hasPendingGenerate } from '@/sequences/ui/generation/pending-generate';
 import { isWelcomeCardAlreadyClaimedError } from '@/platform/errors';
-import { phoneCountries } from '@/ui/phone-countries';
 import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
 import { useNavigate, useSearch } from '@tanstack/react-router';
 import { Sparkles } from 'lucide-react';
@@ -160,8 +155,6 @@ export const WelcomeCreditsProvider: React.FC<{ children: ReactNode }> = ({
   const welcomeSetup = search?.welcome_setup;
   const {
     stripeEnabled,
-    phoneVerificationEnabled,
-    phoneCountry,
     hasUsedCredits,
     hasSignupGrant,
     hasOtherCredits,
@@ -172,7 +165,6 @@ export const WelcomeCreditsProvider: React.FC<{ children: ReactNode }> = ({
   const [skippedUserId, setSkippedUserId] = useState<string | null>(null);
   const [forcedOpen, setForcedOpen] = useState(false);
   const [redirectingToStripe, setRedirectingToStripe] = useState(false);
-  const [viaPhone, setViaPhone] = useState(false);
   const isClient = useSyncExternalStore(
     subscribeNever,
     () => true,
@@ -234,7 +226,6 @@ export const WelcomeCreditsProvider: React.FC<{ children: ReactNode }> = ({
       }
       setForcedOpen(false);
       setSetupError(null);
-      setViaPhone(false);
       if (welcomeSetup) clearWelcomeSetupSearch();
     }
   };
@@ -323,22 +314,6 @@ export const WelcomeCreditsProvider: React.FC<{ children: ReactNode }> = ({
               onStart={() => handleOpenChange(false)}
               primaryLabel={primaryLabel}
             />
-          ) : viaPhone ? (
-            <PhoneClaimDialogContent
-              grantDisplay={GRANT_DISPLAY}
-              showCosts={showCosts}
-              onShowCostsChange={setShowCosts}
-              onGranted={async () => {
-                await queryClient.invalidateQueries({
-                  queryKey: [...BILLING_BALANCE_KEY],
-                });
-                await queryClient.invalidateQueries({
-                  queryKey: [...BILLING_GATE_KEY],
-                });
-              }}
-              onUseCard={() => setViaPhone(false)}
-              defaultCountry={phoneCountry}
-            />
           ) : (
             <ClaimDialogContent
               grantDisplay={GRANT_DISPLAY}
@@ -353,9 +328,6 @@ export const WelcomeCreditsProvider: React.FC<{ children: ReactNode }> = ({
                 setupMutation.mutate();
               }}
               onSkip={() => handleOpenChange(false)}
-              onUsePhone={
-                phoneVerificationEnabled ? () => setViaPhone(true) : undefined
-              }
             />
           )}
         </DialogContent>
@@ -373,7 +345,6 @@ function ClaimDialogContent({
   claiming,
   onAddCard,
   onSkip,
-  onUsePhone,
 }: {
   grantDisplay: string;
   showCosts: boolean;
@@ -383,7 +354,6 @@ function ClaimDialogContent({
   claiming: boolean;
   onAddCard: () => void;
   onSkip: () => void;
-  onUsePhone?: () => void;
 }) {
   const busy = opening || claiming;
   return (
@@ -398,16 +368,7 @@ function ClaimDialogContent({
           {claiming ? 'Unlocking…' : opening ? 'Opening…' : 'Add a card'}
         </Button>
 
-        {onUsePhone ? (
-          <Button
-            variant="link"
-            className="self-center text-muted-foreground"
-            onClick={onUsePhone}
-            disabled={busy}
-          >
-            No card? Verify by SMS instead
-          </Button>
-        ) : null}
+        <AskFounderCard variant="link" />
 
         {setupError ? (
           <p role="alert" className="text-xs text-destructive">
@@ -426,162 +387,6 @@ function ClaimDialogContent({
           Skip for now
         </Button>
       </div>
-    </>
-  );
-}
-
-function errorMessage(err: unknown, fallback: string): string {
-  return err instanceof Error ? err.message : fallback;
-}
-
-/** SMS branch of the claim dialog: number → code → grant. Closes itself via
- *  the balance refetch (`hasSignupGrant` flips, `open` goes false). */
-function PhoneClaimDialogContent({
-  grantDisplay,
-  showCosts,
-  onShowCostsChange,
-  onGranted,
-  onUseCard,
-  defaultCountry,
-}: {
-  grantDisplay: string;
-  showCosts: boolean;
-  onShowCostsChange: (value: boolean) => void;
-  onGranted: () => Promise<void>;
-  onUseCard: () => void;
-  defaultCountry: string | null;
-}) {
-  const [phoneNumber, setPhoneNumber] = useState<string | null>(null);
-  const [code, setCode] = useState('');
-  const countries = useMemo(() => phoneCountries(), []);
-
-  const send = useMutation({
-    meta: { inlineError: true },
-    mutationFn: (data: { phoneNumber: string }) =>
-      sendWelcomePhoneCodeFn({ data }),
-    onSuccess: (result) => setPhoneNumber(result.phoneNumber),
-  });
-  const verify = useMutation({
-    meta: { inlineError: true },
-    mutationFn: (data: { phoneNumber: string; code: string }) =>
-      verifyWelcomePhoneCodeFn({ data }),
-    onSuccess: onGranted,
-  });
-
-  const busy = send.isPending || verify.isPending;
-  const error = verify.error
-    ? isWelcomeCardAlreadyClaimedError(verify.error)
-      ? 'This number has already been used to claim welcome credits'
-      : errorMessage(verify.error, 'Could not verify that code')
-    : send.error
-      ? errorMessage(send.error, 'Could not send a code')
-      : null;
-
-  const onSubmit = (e: React.FormEvent<HTMLFormElement>) => {
-    e.preventDefault();
-    const form = new FormData(e.currentTarget);
-    const field = (name: string) => {
-      const value = form.get(name);
-      return typeof value === 'string' ? value : '';
-    };
-    if (phoneNumber) {
-      verify.mutate({ phoneNumber, code });
-      return;
-    }
-    send.mutate({ phoneNumber: field('phoneNumber') });
-  };
-
-  return (
-    <>
-      <WelcomeHeader
-        amount={grantDisplay}
-        description={
-          phoneNumber
-            ? `Enter the code we texted to ${phoneNumber}.`
-            : "Verify your mobile number to unlock it. We only text you a code — it just confirms you're a real person."
-        }
-      />
-
-      <form onSubmit={onSubmit} className="flex flex-col gap-4 px-6 py-5">
-        {phoneNumber ? (
-          // Plain field, not the slot widget, and NOT autofocused: code
-          // autofill from Messages only offers itself on an unfocused
-          // one-time-code input.
-          <Input
-            name="code"
-            type="text"
-            inputMode="numeric"
-            pattern="[0-9]*"
-            maxLength={6}
-            autoComplete="one-time-code"
-            aria-label="Verification code"
-            placeholder="Enter code"
-            className="text-center tracking-[0.3em] tabular-nums"
-            value={code}
-            onChange={(e) => setCode(e.target.value.replace(/\D/g, ''))}
-            disabled={busy}
-          />
-        ) : (
-          <PhoneInput
-            countries={countries}
-            defaultCountry={defaultCountry}
-            aria-label="Mobile number"
-            disabled={busy}
-            required
-          />
-        )}
-
-        <Button
-          type="submit"
-          className="self-center"
-          disabled={busy || (phoneNumber !== null && code.length !== 6)}
-        >
-          {phoneNumber
-            ? verify.isPending
-              ? 'Unlocking…'
-              : 'Unlock'
-            : send.isPending
-              ? 'Sending…'
-              : 'Send code'}
-        </Button>
-
-        {error ? (
-          <p role="alert" className="text-xs text-destructive">
-            {error}
-          </p>
-        ) : null}
-
-        <ShowCostsRow checked={showCosts} onCheckedChange={onShowCostsChange} />
-
-        <div className="flex justify-between">
-          {phoneNumber ? (
-            <Button
-              type="button"
-              variant="link"
-              className="text-muted-foreground"
-              onClick={() => {
-                setPhoneNumber(null);
-                setCode('');
-                verify.reset();
-              }}
-              disabled={busy}
-            >
-              Change number
-            </Button>
-          ) : (
-            <span />
-          )}
-          <Button
-            type="button"
-            variant="link"
-            className="text-muted-foreground"
-            onClick={onUseCard}
-            disabled={busy}
-          >
-            Use a card instead
-          </Button>
-        </div>
-      </form>
     </>
   );
 }
