+++ src/sequences/pipeline.test.ts
@@ -15,6 +15,7 @@ import {
   flagsFromStopAt,
   GENERATION_STAGES,
   includesStage,
+  allowsUnfundedGeneration,
   isGenerationStage,
   nextActionFromArtifacts,
   nextStageAfter,
@@ -106,6 +107,14 @@ describe('generation pipeline stages', () => {
     expect(includesStage('references', 'references')).toBe(true);
     expect(includesStage('references', 'images')).toBe(false);
     expect(stagesUpTo('references')).toEqual(['script', 'references']);
+  });
+
+  it('lets unfunded teams run through references, not images or motion', () => {
+    expect(allowsUnfundedGeneration('script')).toBe(true);
+    expect(allowsUnfundedGeneration('references')).toBe(true);
+    expect(allowsUnfundedGeneration('images')).toBe(false);
+    expect(allowsUnfundedGeneration('motion')).toBe(false);
+    expect(allowsUnfundedGeneration('music')).toBe(false);
   });
 
   it('runs only the slice from startFrom through stopAt', () => {
