+++ src/sequences/sequences.fn.ts
@@ -52,11 +52,13 @@ import { triggerWorkflow } from '@/platform/server/workflow/client';
 import { triggerStoryboard } from '@/sequences/server/launchers';
 import { ValidationError } from '@/platform/errors';
 import {
+  allowsUnfundedGeneration,
   continueStageSchema,
   flagsFromStopAt,
   generationStageSchema,
   includesStage,
   nextStageAfter,
+  resolveStopAt,
   stageIndex,
 } from './pipeline';
 import type {
@@ -253,33 +255,37 @@ export const continueGenerationFn = createServerFn({ method: 'POST' })
       data.stopAt
     );
 
-    const reservationId = await reserveRunCredits(
-      context.scopedDb,
-      estimateStoryboardPreflightCost({
-        script: sequence.script ?? '',
-        imageModel: safeTextToImageModel(
-          sequence.imageModel,
-          DEFAULT_IMAGE_MODEL
-        ),
-        aspectRatio: sequence.aspectRatio,
-        resolution: sequence.resolution,
-        autoGenerateMotion,
-        stopAt: data.stopAt,
-        startFrom: data.startFrom,
-        referenceOnly: !sequence.generateStartFrames,
-        videoModels: [
-          safeImageToVideoModel(sequence.videoModel, DEFAULT_VIDEO_MODEL),
-        ],
-        autoGenerateMusic,
-        audioModels: [safeAudioModel(sequence.musicModel, DEFAULT_MUSIC_MODEL)],
-        pricing: await getEffectiveFalPricing(),
-      }),
-      {
-        providers: ['fal', 'openrouter'],
-        errorMessage: 'Insufficient credits to continue generation',
-        sequenceId: data.sequenceId,
-      }
-    );
+    const reservationId = allowsUnfundedGeneration(data.stopAt)
+      ? undefined
+      : await reserveRunCredits(
+          context.scopedDb,
+          estimateStoryboardPreflightCost({
+            script: sequence.script ?? '',
+            imageModel: safeTextToImageModel(
+              sequence.imageModel,
+              DEFAULT_IMAGE_MODEL
+            ),
+            aspectRatio: sequence.aspectRatio,
+            resolution: sequence.resolution,
+            autoGenerateMotion,
+            stopAt: data.stopAt,
+            startFrom: data.startFrom,
+            referenceOnly: !sequence.generateStartFrames,
+            videoModels: [
+              safeImageToVideoModel(sequence.videoModel, DEFAULT_VIDEO_MODEL),
+            ],
+            autoGenerateMusic,
+            audioModels: [
+              safeAudioModel(sequence.musicModel, DEFAULT_MUSIC_MODEL),
+            ],
+            pricing: await getEffectiveFalPricing(),
+          }),
+          {
+            providers: ['fal', 'openrouter'],
+            errorMessage: 'Insufficient credits to continue generation',
+            sequenceId: data.sequenceId,
+          }
+        );
 
     await context.scopedDb.sequences.update({
       id: data.sequenceId,
@@ -377,34 +383,41 @@ export const updateSequenceFn = createServerFn({ method: 'POST' })
     }
 
     if (needsRegeneration) {
-      const reservationId = await reserveRunCredits(
-        context.scopedDb,
-        estimateStoryboardPreflightCost({
-          script: sequence.script ?? '',
-          imageModel: safeTextToImageModel(
-            sequence.imageModel,
-            DEFAULT_IMAGE_MODEL
-          ),
-          aspectRatio: sequence.aspectRatio,
-          resolution: sequence.resolution,
-          autoGenerateMotion: sequence.autoGenerateMotion,
-          stopAt: sequence.generationStopAt ?? undefined,
-          videoModels: [
-            safeImageToVideoModel(sequence.videoModel, DEFAULT_VIDEO_MODEL),
-          ],
-          autoGenerateMusic: sequence.autoGenerateMusic,
-          audioModels: [
-            safeAudioModel(sequence.musicModel, DEFAULT_MUSIC_MODEL),
-          ],
-          referenceOnly: !sequence.generateStartFrames,
-          pricing: await getEffectiveFalPricing(),
-        }),
-        {
-          providers: ['fal', 'openrouter'],
-          errorMessage: 'Insufficient credits to regenerate storyboard',
-          sequenceId,
-        }
-      );
+      const stopAt = resolveStopAt({
+        generationStopAt: sequence.generationStopAt,
+        autoGenerateMotion: sequence.autoGenerateMotion,
+        autoGenerateMusic: sequence.autoGenerateMusic,
+      });
+      const reservationId = allowsUnfundedGeneration(stopAt)
+        ? undefined
+        : await reserveRunCredits(
+            context.scopedDb,
+            estimateStoryboardPreflightCost({
+              script: sequence.script ?? '',
+              imageModel: safeTextToImageModel(
+                sequence.imageModel,
+                DEFAULT_IMAGE_MODEL
+              ),
+              aspectRatio: sequence.aspectRatio,
+              resolution: sequence.resolution,
+              autoGenerateMotion: sequence.autoGenerateMotion,
+              stopAt,
+              videoModels: [
+                safeImageToVideoModel(sequence.videoModel, DEFAULT_VIDEO_MODEL),
+              ],
+              autoGenerateMusic: sequence.autoGenerateMusic,
+              audioModels: [
+                safeAudioModel(sequence.musicModel, DEFAULT_MUSIC_MODEL),
+              ],
+              referenceOnly: !sequence.generateStartFrames,
+              pricing: await getEffectiveFalPricing(),
+            }),
+            {
+              providers: ['fal', 'openrouter'],
+              errorMessage: 'Insufficient credits to regenerate storyboard',
+              sequenceId,
+            }
+          );
 
       // Owns the generation mutex, the 'processing' status write, the run-id
       // persistence (#839), and the trigger-time content snapshot. Regeneration
@@ -522,31 +535,41 @@ export const retryStoryboardFn = createServerFn({ method: 'POST' })
       throw new Error('Only failed sequences can be retried');
     }
 
-    const reservationId = await reserveRunCredits(
-      context.scopedDb,
-      estimateStoryboardPreflightCost({
-        script: sequence.script ?? '',
-        imageModel: safeTextToImageModel(
-          sequence.imageModel,
-          DEFAULT_IMAGE_MODEL
-        ),
-        aspectRatio: sequence.aspectRatio,
-        resolution: sequence.resolution,
-        autoGenerateMotion: sequence.autoGenerateMotion,
-        videoModels: [
-          safeImageToVideoModel(sequence.videoModel, DEFAULT_VIDEO_MODEL),
-        ],
-        autoGenerateMusic: sequence.autoGenerateMusic,
-        audioModels: [safeAudioModel(sequence.musicModel, DEFAULT_MUSIC_MODEL)],
-        referenceOnly: !sequence.generateStartFrames,
-        pricing: await getEffectiveFalPricing(),
-      }),
-      {
-        providers: ['fal', 'openrouter'],
-        errorMessage: 'Insufficient credits to retry storyboard',
-        sequenceId: sequence.id,
-      }
-    );
+    const stopAt = resolveStopAt({
+      generationStopAt: sequence.generationStopAt,
+      autoGenerateMotion: sequence.autoGenerateMotion,
+      autoGenerateMusic: sequence.autoGenerateMusic,
+    });
+    const reservationId = allowsUnfundedGeneration(stopAt)
+      ? undefined
+      : await reserveRunCredits(
+          context.scopedDb,
+          estimateStoryboardPreflightCost({
+            script: sequence.script ?? '',
+            imageModel: safeTextToImageModel(
+              sequence.imageModel,
+              DEFAULT_IMAGE_MODEL
+            ),
+            aspectRatio: sequence.aspectRatio,
+            resolution: sequence.resolution,
+            autoGenerateMotion: sequence.autoGenerateMotion,
+            stopAt,
+            videoModels: [
+              safeImageToVideoModel(sequence.videoModel, DEFAULT_VIDEO_MODEL),
+            ],
+            autoGenerateMusic: sequence.autoGenerateMusic,
+            audioModels: [
+              safeAudioModel(sequence.musicModel, DEFAULT_MUSIC_MODEL),
+            ],
+            referenceOnly: !sequence.generateStartFrames,
+            pricing: await getEffectiveFalPricing(),
+          }),
+          {
+            providers: ['fal', 'openrouter'],
+            errorMessage: 'Insufficient credits to retry storyboard',
+            sequenceId: sequence.id,
+          }
+        );
 
     const workflowInput: StoryboardTriggerInput = {
       userId: user.id,
@@ -562,7 +585,7 @@ export const retryStoryboardFn = createServerFn({ method: 'POST' })
       },
       autoGenerateMotion: sequence.autoGenerateMotion,
       autoGenerateMusic: sequence.autoGenerateMusic,
-      stopAt: sequence.generationStopAt ?? undefined,
+      stopAt,
     };
 
     // Owns the generation mutex, the 'processing' status write, and the
