+++ src/sequences/ui/script/script-view.tsx
@@ -46,6 +46,7 @@ import { BILLING_TRANSACTIONS_KEY } from '@/billing/ui/use-billing-balance-realt
 import { useBillingGate } from '@/billing/ui/use-billing-gate';
 import { useGenerationSettings } from '@/sequences/ui/use-generation-settings';
 import {
+  allowsUnfundedGeneration,
   DEFAULT_GENERATION_STOP_AT,
   flagsFromStopAt,
   includesStage,
@@ -895,6 +896,10 @@ export const ScriptView: FC<{
     > = genSettings
   ) => {
     const { stopAt: runUntil, generateStartFrames, videoModels } = run;
+    if (needsBillingSetup && !allowsUnfundedGeneration(runUntil)) {
+      showGate();
+      return;
+    }
     const flags = flagsFromStopAt(runUntil);
     // sequence_generated is captured server-side in createSequences (#1088)
     // so dashboard + public API both feed #product-alerts once.
@@ -946,7 +951,12 @@ export const ScriptView: FC<{
   };
 
   const requestGenerate = () => {
-    if (savedSettings.rememberStopAt) {
+    // Remembered paid stops still open the slider when the team has no
+    // credits, so they can slide back to References (#1566).
+    if (
+      savedSettings.rememberStopAt &&
+      !(needsBillingSetup && !allowsUnfundedGeneration(stopAt))
+    ) {
       executeRegeneration();
       return;
     }
@@ -991,11 +1001,6 @@ export const ScriptView: FC<{
       return;
     }
 
-    if (needsBillingSetup) {
-      showGate();
-      return;
-    }
-
     if (isEditing) {
       if (savedSettings.rememberStopAt) {
         setEnhance('showRegenerateConfirm', true);
@@ -1034,11 +1039,6 @@ export const ScriptView: FC<{
     // (#1286).
     if (!requireAuth()) {
       if (!isEditing) markPendingIntent('enhance');
-      return;
-    }
-
-    if (needsBillingSetup) {
-      showGate();
       return;
     }
